backend/requirements.txt
django
djangorestframework
django-cors-headers
tzdata
pillow